0.5.0 release

+++ app/dance/spinners/circle.go
@@ -7,17 +7,17 @@ import (
 )
 
 type CircleMover struct {
-	start int64
+	start float64
 }
 
 func NewCircleMover() *CircleMover {
 	return &CircleMover{}
 }
 
-func (c *CircleMover) Init(start, end int64) {
+func (c *CircleMover) Init(start, end float64) {
 	c.start = start
 }
 
-func (c *CircleMover) GetPositionAt(time int64) vector.Vector2f {
+func (c *CircleMover) GetPositionAt(time float64) vector.Vector2f {
 	return vector.NewVec2fRad(rpms*float32(time-c.start)*2*math32.Pi, float32(settings.Dance.SpinnerRadius)).Add(center)
 }
